requirements.txt
-i https://pypi.org/simple
asgiref==3.11.1; python_version >= '3.9'
django==6.0.4; python_version >= '3.12'
djangorestframework==3.17.1; python_version >= '3.10'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
pillow==12.2.0; python_version >= '3.10'
pyjwt==2.12.1; python_version >= '3.9'
sqlparse==0.5.5; python_version >= '3.8'
